Port detailed query page to Typescript
