+++ site/frontend/src/urls.ts
@@ -1,4 +1,4 @@
-export const BASE_URL = window.location.origin + "/perf";
+const BASE_URL = window.location.origin + "/perf";
 
 export const INFO_URL = `${BASE_URL}/info`;
 
@@ -7,3 +7,4 @@ export const STATUS_DATA_URL = `${BASE_URL}/status_page`;
 export const BOOTSTRAP_DATA_URL = `${BASE_URL}/bootstrap`;
 export const GRAPH_DATA_URL = `${BASE_URL}/graphs`;
 export const COMPARE_DATA_URL = `${BASE_URL}/get`;
+export const SELF_PROFILE_DATA_URL = `${BASE_URL}/self-profile`;
