+++ site/frontend/templates/pages/detailed-query.html
@@ -103,7 +103,5 @@
 </div>
 {% endblock content %}
 {% block script %}
-<script src="https://cdnjs.cloudflare.com/ajax/libs/msgpack-lite/0.1.26/msgpack.min.js"></script>
-<script src="scripts/shared.js"></script>
 <script src="scripts/detailed-query.js"></script>
 {% endblock script %}
